+++ crypto/tls/tls_test.go
@@ -1,9 +1,8 @@
-package kvtls
+package tls
 
 import (
-	"fmt"
-	"io/ioutil"
 	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -70,32 +69,28 @@ type x509Paths struct {
 }
 
 func newTestX509Files(t *testing.T, cert, key, ca []byte) x509Paths {
-	// create empty file
-	certsPath, err := ioutil.TempDir("", "*-x509")
-	require.NoError(t, err)
+	t.Helper()
 
-	t.Cleanup(func() {
-		os.RemoveAll(certsPath)
-	})
+	certsPath := t.TempDir()
 
 	paths := x509Paths{
-		cert: fmt.Sprintf("%s/cert.pem", certsPath),
-		key:  fmt.Sprintf("%s/key.pem", certsPath),
-		ca:   fmt.Sprintf("%s/ca.pem", certsPath),
+		cert: filepath.Join(certsPath, "cert.pem"),
+		key:  filepath.Join(certsPath, "key.pem"),
+		ca:   filepath.Join(certsPath, "ca.pem"),
 	}
 
 	if cert != nil {
-		err = ioutil.WriteFile(paths.cert, cert, 0600)
+		err := os.WriteFile(paths.cert, cert, 0600)
 		require.NoError(t, err)
 	}
 
 	if key != nil {
-		err = ioutil.WriteFile(paths.key, key, 0600)
+		err := os.WriteFile(paths.key, key, 0600)
 		require.NoError(t, err)
 	}
 
 	if ca != nil {
-		err = ioutil.WriteFile(paths.ca, ca, 0600)
+		err := os.WriteFile(paths.ca, ca, 0600)
 		require.NoError(t, err)
 	}
