+++ kv/memberlist/tcp_transport.go
@@ -21,8 +21,8 @@ import (
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"go.uber.org/atomic"
 
+	dstls "github.com/grafana/dskit/crypto/tls"
 	"github.com/grafana/dskit/flagext"
-	"github.com/grafana/dskit/kv/kvtls"
 )
 
 type messageType uint8
@@ -58,7 +58,7 @@ type TCPTransportConfig struct {
 	MetricsNamespace  string                `yaml:"-"`
 
 	TLSEnabled bool               `yaml:"tls_enabled"`
-	TLS        kvtls.ClientConfig `yaml:",inline"`
+	TLS        dstls.ClientConfig `yaml:",inline"`
 }
 
 // RegisterFlags registers flags.
